% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets_the_range.R
\name{get_range}
\alias{get_range}
\title{Gets the range of at least 6 data points}
\usage{
get_range(mynumber)
}
\arguments{
\item{mynumber}{is any vector of numbers}
}
\value{
a dataframe with upper and lower data
}
\description{
Gets the range of at least 6 data points
}
