% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot_model_1_label.R
\name{myplot_model_1_label}
\alias{myplot_model_1_label}
\title{myplot_model_1_label}
\usage{
myplot_model_1_label(filename, vardep, varindep, desiredR2 = 0.8)
}
\description{
myplot_model_1_label
}
