% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_data_discriminatory_concentrations.R
\name{reading_data_discriminatory_concentration}
\alias{reading_data_discriminatory_concentration}
\title{reading_data_discriminatory_concentrations}
\usage{
reading_data_discriminatory_concentration(filename, length_repeats, plug)
}
\description{
reading_data_discriminatory_concentrations
}
