% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_serial_dilution.R
\name{reading_data_serial_dilution}
\alias{reading_data_serial_dilution}
\title{Reading serial dilution entry}
\usage{
reading_data_serial_dilution(filename, length_repeats, plug)
}
\description{
Reading serial dilution entry
}
